Merge develop into master 1.0.5

+++ tools/istat.c
@@ -41,8 +41,8 @@ int main(int argc,  char** argv)
     }
 
     if (INA_SUCCEED(ina_dir_stat_new(ina_str_cstr(dir), &__ds))) {
-        uint64_t b_total;
-        uint64_t b_free;
+        size_t b_total;
+        size_t b_free;
         ina_dir_stat_bytes_capacity(__ds, &b_total);
         ina_dir_stat_bytes_free(__ds, &b_free);
         if (hr) {
